refactor(bottlecap): promote start_trace_agent to library crate
Moves the start_trace_agent helper from the private bin/bottlecap/main.rs
into a new public module, bottlecap::traces::startup, so callers outside
the Lambda binary (notably the forthcoming bottlecap-testmode binary)
can construct the full trace-processing pipeline through a single call
instead of duplicating ~110 lines of wiring.

No behavior change. The Lambda binary's call site is unchanged: same
function name, same signature, same internal spawn.
